+++ src/rer/ufficiostampa/restapi/services/search_comunicati/get.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+from rer.ufficiostampa.interfaces import IRerUfficiostampaSettings
+from plone.restapi.services.search.get import SearchGet
+
+
+class SearchComunicatiGet(SearchGet):
+    """ """